// src/api/generated/models/getAllUsersApiV1UsersGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { LifecycleStatusFilter } from './lifecycleStatusFilter'
import type { Role } from './role'
import type { SortOrder } from './sortOrder'
import type { UserSortField } from './userSortField'

export type GetAllUsersApiV1UsersGetParams = {
  office_id?: string | null
  role?: Role | null
  status?: LifecycleStatusFilter
  sort_by?: UserSortField
  order?: SortOrder
  /**
   * @minimum 1
   */
  page?: number
  /**
   * @minimum 1
   * @maximum 100
   */
  size?: number
  q?: string | null
}