// src/api/generated/models/role.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Define the supported citizen and authority user roles.
 */
export type Role = (typeof Role)[keyof typeof Role]

export const Role = {
  CITIZEN: 'CITIZEN',
  DISPATCHER: 'DISPATCHER',
  OFFICER: 'OFFICER',
  MANAGER: 'MANAGER',
  ADMIN: 'ADMIN',
} as const