// src/api/generated/models/sortOrder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Define ascending and descending list ordering values.
 */
export type SortOrder = (typeof SortOrder)[keyof typeof SortOrder]

export const SortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const