// src/api/generated/models/userSortField.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */

/**
 * Define supported sort fields for user list queries.
 */
export type UserSortField = (typeof UserSortField)[keyof typeof UserSortField]

export const UserSortField = {
  created_at: 'created_at',
  email: 'email',
  first_name: 'first_name',
  last_name: 'last_name',
  role: 'role',
} as const